chore: arrange ui module packages
