+++ core/ui/src/main/java/com/emotionstorage/ui/component/bottomSheet/DatePickerBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.ui.component
+package com.emotionstorage.ui.component.bottomSheet
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -186,11 +186,12 @@ fun DatePickerBottomSheet(
                                         .clickable(
                                             enabled = date.isAfter(minDate) && date.isBefore(maxDate),
                                             onClick = {
-                                                scope.launch { sheetState.hide() }.invokeOnCompletion {
-                                                    if (!sheetState.isVisible) {
-                                                        onDateSelect(date)
+                                                scope.launch { sheetState.hide() }
+                                                    .invokeOnCompletion {
+                                                        if (!sheetState.isVisible) {
+                                                            onDateSelect(date)
+                                                        }
                                                     }
-                                                }
                                             },
                                         ),
                             )
