+++ core/ui/src/main/java/com/emotionstorage/ui/component/picker/TimeWheelSpinner.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.ui.component
+package com.emotionstorage.ui.component.picker
 
 import androidx.compose.animation.animateColorAsState
 import androidx.compose.foundation.background
