+++ core/ui/src/main/java/com/emotionstorage/ui/component/text/RichTextList.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.ui.component
+package com.emotionstorage.ui.component.text
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -7,6 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.emotionstorage.ui.content.ListStyle
 import com.emotionstorage.ui.theme.MooiTheme
 
 private const val WS = "[\\s\\u00A0]"
