+++ core/ui/src/main/java/com/emotionstorage/ui/component/text/TextBoxInput.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.ui.component
+package com.emotionstorage.ui.component.text
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
