+++ core/ui/src/main/java/com/emotionstorage/ui/content/TermsOfServiceContent.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.ui.component
+package com.emotionstorage.ui.content
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Spacer
@@ -19,6 +19,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.emotionstorage.ui.theme.MooiTheme
 import com.emotionstorage.ui.R
+import com.emotionstorage.ui.component.text.RichBody
+import com.emotionstorage.ui.component.text.RichList
 
 enum class ListStyle { Bulleted, Numbered }
