+++ feat/my/src/main/java/com/emotionstorage/my/ui/TermsAndPrivacyScreen.kt
@@ -22,8 +22,8 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.emotionstorage.ui.component.PrivacyPolicyContent
-import com.emotionstorage.ui.component.TermsOfServiceContent
+import com.emotionstorage.ui.content.PrivacyPolicyContent
+import com.emotionstorage.ui.content.TermsOfServiceContent
 import com.emotionstorage.ui.component.TopAppBar
 import com.emotionstorage.ui.theme.MooiTheme
