+++ feat/time-capsule-detail/src/main/java/com/emotionstorage/time_capsule_detail/ui/component/SaveTimeCapsuleButton.kt
@@ -18,7 +18,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.emotionstorage.ui.component.CountDownTimer
-import com.emotionstorage.ui.component.CtaButton
+import com.emotionstorage.ui.component.button.CtaButton
 import com.emotionstorage.ui.theme.MooiTheme
 import java.time.LocalDateTime
