+++ feat/time-capsule-detail/src/main/java/com/emotionstorage/time_capsule_detail/ui/component/TimeCapsuleNote.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.emotionstorage.ui.component.TextBoxInput
+import com.emotionstorage.ui.component.text.TextBoxInput
 import com.emotionstorage.ui.theme.MooiTheme
 import com.orhanobut.logger.Logger
 import kotlinx.coroutines.Job
