+++ feat/time-capsule/src/main/java/com/emotionstorage/time_capsule/ui/FavoriteTimeCapsulesScreen.kt
@@ -42,9 +42,9 @@ import com.emotionstorage.time_capsule.ui.model.TimeCapsuleItemState
 import com.emotionstorage.ui.component.TopAppBar
 import com.emotionstorage.ui.theme.MooiTheme
 import com.emotionstorage.ui.R
-import com.emotionstorage.ui.component.AppSnackbarHost
-import com.emotionstorage.ui.component.DropDownPicker
-import com.emotionstorage.ui.component.FavoriteToast
+import com.emotionstorage.ui.component.toast.AppSnackbarHost
+import com.emotionstorage.ui.component.picker.DropDownPicker
+import com.emotionstorage.ui.component.toast.FavoriteToast
 import java.time.LocalDateTime
 
 @Composable
