+++ feat/tutorial/src/main/java/com/emotionstorage/tutorial/ui/onBoarding/AgreeTermsScreen.kt
@@ -33,8 +33,8 @@ import com.emotionstorage.tutorial.R
 import com.emotionstorage.tutorial.presentation.onBoarding.AgreeTermsEvent
 import com.emotionstorage.tutorial.presentation.onBoarding.AgreeTermsViewModel
 import com.emotionstorage.tutorial.presentation.onBoarding.AgreeTermsViewModel.State
-import com.emotionstorage.ui.component.CtaButton
-import com.emotionstorage.ui.component.ToggleButton
+import com.emotionstorage.ui.component.button.CtaButton
+import com.emotionstorage.ui.component.button.ToggleButton
 import com.emotionstorage.ui.component.TopAppBar
 import com.emotionstorage.ui.theme.MooiTheme
 import kotlinx.coroutines.launch
